Moving tests to their own unit
