+++ tests/unit/all.ts
@@ -4,6 +4,7 @@ import './WidgetRegistry';
 import './lifecycle';
 import './customElements';
 import './d';
+import './decorators/all';
 import './mixins/all';
 import './util/all';
 import './main';
