+++ tests/unit/decorators/all.ts
@@ -1,0 +1,1 @@
+import './propagateProperty';
